Update ID type from Long to BigDecimal for DB2 Decimal compatibility

+++ src/main/java/com/example/demo/service/UserService.java
@@ -1,12 +1,13 @@
 package com.example.demo.service;
 
 import com.example.demo.model.User;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface UserService {
     List<User> getAllUsers();
-    User getUserById(Long id);
+    User getUserById(BigDecimal id);
     User createUser(User user);
-    User updateUser(Long id, User user);
-    void deleteUser(Long id);
+    User updateUser(BigDecimal id, User user);
+    void deleteUser(BigDecimal id);
 }
